+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `insert` and `extract` instructions.
 
 ### Changed
+- Rename `VectorType` to `ArrayType`. ([#24](https://github.com/fabianschuiki/llhd/issues/24))
+- Change vector syntax from `<N x T>` to `[N x T]`.
 - Change type of delta and epsilon time steps to `usize`.
 
 ### Fixed